sources: refactor bazaar source into module.
Extract the Bazaar source into `sources/_bazaar.py`.

LP: #1648906
