+++ snapcraft/internal/sources/__init__.py
@@ -89,6 +89,8 @@ from snapcraft import file_utils
 from . import errors
 from . import _base
 
+from ._bazaar import Bazaar  # noqa
+
 
 logging.getLogger('urllib3').setLevel(logging.CRITICAL)
 
@@ -119,40 +121,6 @@ class Script(_base.FileBase):
         super().download()
         st = os.stat(self.file)
         os.chmod(self.file, st.st_mode | stat.S_IEXEC)
-
-
-class Bazaar(_base.Base):
-
-    def __init__(self, source, source_dir, source_tag=None, source_commit=None,
-                 source_branch=None, source_depth=None):
-        super().__init__(source, source_dir, source_tag, source_commit,
-                         source_branch, source_depth, 'bzr')
-        if source_branch:
-            raise errors.IncompatibleOptionsError(
-                'can\'t specify a source-branch for a bzr source')
-        if source_depth:
-            raise errors.IncompatibleOptionsError(
-                'can\'t specify source-depth for a bzr source')
-        if source_tag and source_commit:
-            raise errors.IncompatibleOptionsError(
-                'can\'t specify both source-tag and source-commit for '
-                'a bzr source')
-
-    def pull(self):
-        tag_opts = []
-        if self.source_tag:
-            tag_opts = ['-r', 'tag:' + self.source_tag]
-        if self.source_commit:
-            tag_opts = ['-r', self.source_commit]
-        if os.path.exists(os.path.join(self.source_dir, '.bzr')):
-            cmd = [self.command, 'pull'] + tag_opts + \
-                  [self.source, '-d', self.source_dir]
-        else:
-            os.rmdir(self.source_dir)
-            cmd = [self.command, 'branch'] + tag_opts + \
-                  [self.source, self.source_dir]
-
-        subprocess.check_call(cmd)
 
 
 class Git(_base.Base):
