+++ snapcraft/tests/sources/test_sources.py
@@ -248,89 +248,6 @@ class TestRpm(tests.TestCase):
         self.assertCountEqual(os.listdir(dest_dir), test_output_files)
 
 
-class TestBazaar(SourceTestCase):
-
-    def test_pull(self):
-        bzr = sources.Bazaar('lp:my-source', 'source_dir')
-
-        bzr.pull()
-
-        self.mock_rmdir.assert_called_once_with('source_dir')
-        self.mock_run.assert_called_once_with(
-            ['bzr', 'branch', 'lp:my-source', 'source_dir'])
-
-    def test_pull_tag(self):
-        bzr = sources.Bazaar(
-            'lp:my-source', 'source_dir', source_tag='tag')
-        bzr.pull()
-
-        self.mock_run.assert_called_once_with(
-            ['bzr', 'branch', '-r', 'tag:tag', 'lp:my-source',
-             'source_dir'])
-
-    def test_pull_existing_with_tag(self):
-        self.mock_path_exists.return_value = True
-
-        bzr = sources.Bazaar(
-            'lp:my-source', 'source_dir', source_tag='tag')
-        bzr.pull()
-
-        self.mock_run.assert_called_once_with(
-            ['bzr', 'pull', '-r', 'tag:tag', 'lp:my-source', '-d',
-             'source_dir'])
-
-    def test_pull_commit(self):
-        bzr = sources.Bazaar(
-            'lp:my-source', 'source_dir', source_commit='2')
-        bzr.pull()
-
-        self.mock_run.assert_called_once_with(
-            ['bzr', 'branch', '-r', '2', 'lp:my-source',
-             'source_dir'])
-
-    def test_pull_existing_with_commit(self):
-        self.mock_path_exists.return_value = True
-
-        bzr = sources.Bazaar(
-            'lp:my-source', 'source_dir', source_commit='2')
-        bzr.pull()
-
-        self.mock_run.assert_called_once_with(
-            ['bzr', 'pull', '-r', '2', 'lp:my-source', '-d',
-             'source_dir'])
-
-    def test_init_with_source_branch_raises_exception(self):
-        raised = self.assertRaises(
-            sources.errors.IncompatibleOptionsError,
-            sources.Bazaar,
-            'lp:mysource', 'source_dir', source_branch='branch')
-
-        expected_message = 'can\'t specify a source-branch for a bzr source'
-        self.assertEqual(raised.message, expected_message)
-
-    def test_init_with_source_depth_raises_exception(self):
-        raised = self.assertRaises(
-            sources.errors.IncompatibleOptionsError,
-            sources.Bazaar,
-            'lp://mysource', 'source_dir', source_depth=2)
-
-        expected_message = (
-            'can\'t specify source-depth for a bzr source')
-        self.assertEqual(raised.message, expected_message)
-
-    def test_init_with_source_tag_and_commit_raises_exception(self):
-        raised = self.assertRaises(
-            sources.errors.IncompatibleOptionsError,
-            sources.Bazaar,
-            'lp://mysource', 'source_dir', source_tag="tag",
-            source_commit="2")
-
-        expected_message = (
-            'can\'t specify both source-tag and source-commit for '
-            'a bzr source')
-        self.assertEqual(raised.message, expected_message)
-
-
 class TestGit(SourceTestCase):
 
     def test_pull(self):
